Delete file only if file exists

+++ app/mailers/message_mailer.rb
@@ -3,6 +3,6 @@ class MessageMailer < ApplicationMailer
     @body = body
     attachments["post_event.ics"] = File.read(file_name) if file_name
     mail to: recipient, subject: "foodstream: #{subject}", message: body
-    File.delete(file_name)
+    File.delete(file_name) if file_name
   end
 end
